Rest_API_Django_example/coding-challenge-api/requirements.txt
appnope==0.1.0
autopep8==1.3.4
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
colorlover==0.2.1
cufflinks==0.12.1
cycler==0.10.0
decorator==4.2.1
Django==2.0.2
djangorestframework==3.7.7
docopt==0.6.2
entrypoints==0.2.3
et-xmlfile==1.0.1
freeze==1.0.10
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.4
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.4.0
numpy==1.14.0
openpyxl==2.5.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pipreqs==0.4.9
plotly==2.4.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
yarg==0.1.9
